"Add support for automatic GDB invocation upon temporary binary (for debugging).
Just pass a "--debug" parameter to the invoked script.  Though, the
parameter must be the FIRST parameter passed.  It will also be
swallowed, so there is a super slim chance that a script also requires
such a command-line switch and it may be confusing to the user that
they would have to pass it twice.  I'm satisfied that it's never going
to be an issue for me though (future me: if you are here reading this
trying to track down why your --debug isn't making it to your script,
sorry, you should have tried to remember harder).